Add support for watch and add example to stream watch changes.
This adds another example that streams the 'watch' API using Python as a Server
Side Event that the browser is able to display. The events are simply the
stream of events from etcd and are not doing any complicated processing other
than just rendering them in the browser page.

See examples/03-forum-webapp/ on how to run the example.